abstract out common 3D viewer app structure
