+++ portal/src/main/java/com/hellblazer/luciferase/portal/GridInspector.java
@@ -39,28 +39,14 @@ import javax.vecmath.Point3f;
 import com.hellblazer.luciferase.lucien.grid.MutableGrid;
 import com.hellblazer.luciferase.lucien.grid.Vertex;
 
-import javafx.application.Application;
-import javafx.event.EventHandler;
-import javafx.scene.DepthTest;
-import javafx.scene.Group;
-import javafx.scene.Node;
-import javafx.scene.PerspectiveCamera;
-import javafx.scene.Scene;
-import javafx.scene.input.KeyEvent;
-import javafx.scene.input.MouseEvent;
-import javafx.scene.input.ScrollEvent;
-import javafx.scene.paint.Color;
-import javafx.scene.paint.PhongMaterial;
-import javafx.scene.shape.Box;
-import javafx.stage.Stage;
-
 /**
  * Neolithic 3D viewer, based on ye venerable JavaFX 3D sample app
  *
  * @author hal.hildebrand
  * @author cmcastil
  */
-public class GridInspector extends Application {
+public class GridInspector extends Abstract3DApp {
+
     /**
      * This is the main() you want to run from your IDE
      */
@@ -71,78 +57,12 @@ public class GridInspector extends Application {
         }
     }
 
-    private static final double AXIS_LENGTH             = 250.0;
-    private static final double CAMERA_FAR_CLIP         = 10000.0;
-    private static final double CAMERA_INITIAL_DISTANCE = -450;
-    private static final double CAMERA_INITIAL_X_ANGLE  = 70.0;
-    private static final double CAMERA_INITIAL_Y_ANGLE  = 320.0;
-    private static final double CAMERA_NEAR_CLIP        = 0.1;
-    private static final double CONTROL_MULTIPLIER      = 0.1;
-    private static final double MOUSE_SPEED             = 0.1;
-    private static final double ROTATION_SPEED          = 2.0;
-    private static final double SHIFT_MULTIPLIER        = 10.0;
-    private static final double TRACK_SPEED             = 0.3;
-
     public static void main(String[] args) {
         launch(args);
     }
 
-    final Xform             axisGroup         = new Xform();
-    final PerspectiveCamera camera            = new PerspectiveCamera(true);
-    final Xform             cameraXform       = new Xform();
-    final Xform             cameraXform2      = new Xform();
-    final Xform             cameraXform3      = new Xform();
-    double                  mouseDeltaX;
-    double                  mouseDeltaY;
-    double                  mouseOldX;
-    double                  mouseOldY;
-    double                  mousePosX;
-    double                  mousePosY;
-    final Group             root              = new Group();
-    final Xform             transformingGroup = new Xform();
-    final Xform             world             = new Xform();
-    private GridView        view;
-
     @Override
-    public void start(Stage primaryStage) {
-
-        // setUserAgentStylesheet(STYLESHEET_MODENA);
-
-        root.getChildren().add(world);
-        root.setDepthTest(DepthTest.ENABLE);
-
-        // buildScene();
-        buildCamera();
-        buildAxes();
-        build();
-
-        Scene scene = new Scene(root, 1024, 768, true);
-        scene.setFill(Color.LIGHTGRAY);
-        handleKeyboard(scene, world);
-        handleMouse(scene, world);
-
-        primaryStage.setTitle("Sentinel Inspector");
-        primaryStage.setScene(scene);
-        primaryStage.show();
-
-        scene.setCamera(camera);
-        // Attach a scroll listener
-        primaryStage.addEventHandler(ScrollEvent.SCROLL, event -> {
-            double modifier = 50.0;
-            double modifierFactor = 0.01;
-            if (event.isControlDown()) {
-                modifier = 1;
-            }
-            if (event.isShiftDown()) {
-                modifier = 100.0;
-            }
-            double z = camera.getTranslateZ();
-            double newZ = z + event.getDeltaY() * modifierFactor * modifier;
-            camera.setTranslateZ(newZ);
-        });
-    }
-
-    private void build() {
+    protected void build() {
         final var random = new Random(666);
         final var tet = new MutableGrid();
         Point3f ourPoints[] = Vertex.getRandomPoints(random, 200, 10.0f, true);
@@ -154,116 +74,5 @@ public class GridInspector extends Application {
         transformingGroup.getChildren().add(view);
 
         world.getChildren().addAll(transformingGroup);
-    }
-
-    private void buildAxes() {
-        final PhongMaterial redMaterial = new PhongMaterial();
-        redMaterial.setDiffuseColor(Color.DARKRED);
-        redMaterial.setSpecularColor(Color.RED);
-
-        final PhongMaterial greenMaterial = new PhongMaterial();
-        greenMaterial.setDiffuseColor(Color.DARKGREEN);
-        greenMaterial.setSpecularColor(Color.GREEN);
-
-        final PhongMaterial blueMaterial = new PhongMaterial();
-        blueMaterial.setDiffuseColor(Color.DARKBLUE);
-        blueMaterial.setSpecularColor(Color.BLUE);
-
-        final Box xAxis = new Box(AXIS_LENGTH, 1, 1);
-        final Box yAxis = new Box(1, AXIS_LENGTH, 1);
-        final Box zAxis = new Box(1, 1, AXIS_LENGTH);
-
-        xAxis.setMaterial(redMaterial);
-        yAxis.setMaterial(greenMaterial);
-        zAxis.setMaterial(blueMaterial);
-
-        axisGroup.getChildren().addAll(xAxis, yAxis, zAxis);
-        axisGroup.setVisible(false);
-        world.getChildren().addAll(axisGroup);
-    }
-
-    private void buildCamera() {
-        root.getChildren().add(cameraXform);
-        cameraXform.getChildren().add(cameraXform2);
-        cameraXform2.getChildren().add(cameraXform3);
-        cameraXform3.getChildren().add(camera);
-        cameraXform3.setRotateZ(180.0);
-
-        camera.setNearClip(CAMERA_NEAR_CLIP);
-        camera.setFarClip(CAMERA_FAR_CLIP);
-        camera.setTranslateZ(CAMERA_INITIAL_DISTANCE);
-        cameraXform.ry.setAngle(CAMERA_INITIAL_Y_ANGLE);
-        cameraXform.rx.setAngle(CAMERA_INITIAL_X_ANGLE);
-    }
-
-    private void handleKeyboard(Scene scene, final Node root) {
-        scene.setOnKeyPressed(new EventHandler<KeyEvent>() {
-            @Override
-            public void handle(KeyEvent event) {
-                switch (event.getCode()) {
-                case Z:
-                    cameraXform2.t.setX(0.0);
-                    cameraXform2.t.setY(0.0);
-                    camera.setTranslateZ(CAMERA_INITIAL_DISTANCE);
-                    cameraXform.ry.setAngle(CAMERA_INITIAL_Y_ANGLE);
-                    cameraXform.rx.setAngle(CAMERA_INITIAL_X_ANGLE);
-                    break;
-                case X:
-                    axisGroup.setVisible(!axisGroup.isVisible());
-                    break;
-                case V:
-                    transformingGroup.setVisible(!transformingGroup.isVisible());
-                    break;
-                default:
-                    break;
-                }
-            }
-        });
-    }
-
-    private void handleMouse(Scene scene, final Node root) {
-        scene.setOnMousePressed(new EventHandler<MouseEvent>() {
-            @Override
-            public void handle(MouseEvent me) {
-                mousePosX = me.getSceneX();
-                mousePosY = me.getSceneY();
-                mouseOldX = me.getSceneX();
-                mouseOldY = me.getSceneY();
-            }
-        });
-        scene.setOnMouseDragged(new EventHandler<MouseEvent>() {
-            @Override
-            public void handle(MouseEvent me) {
-                mouseOldX = mousePosX;
-                mouseOldY = mousePosY;
-                mousePosX = me.getSceneX();
-                mousePosY = me.getSceneY();
-                mouseDeltaX = (mousePosX - mouseOldX);
-                mouseDeltaY = (mousePosY - mouseOldY);
-
-                double modifier = 1.0;
-
-                if (me.isControlDown()) {
-                    modifier = CONTROL_MULTIPLIER;
-                }
-                if (me.isShiftDown()) {
-                    modifier = SHIFT_MULTIPLIER;
-                }
-
-                if (me.isMiddleButtonDown() || (me.isPrimaryButtonDown() && me.isSecondaryButtonDown())) {
-                    cameraXform2.t.setX(cameraXform2.t.getX() + mouseDeltaX * MOUSE_SPEED * modifier * TRACK_SPEED);
-                    cameraXform2.t.setY(cameraXform2.t.getY() + mouseDeltaY * MOUSE_SPEED * modifier * TRACK_SPEED);
-                } else if (me.isPrimaryButtonDown()) {
-                    cameraXform.ry.setAngle(cameraXform.ry.getAngle()
-                    - mouseDeltaX * MOUSE_SPEED * modifier * ROTATION_SPEED);
-                    cameraXform.rx.setAngle(cameraXform.rx.getAngle()
-                    + mouseDeltaY * MOUSE_SPEED * modifier * ROTATION_SPEED);
-                } else if (me.isSecondaryButtonDown()) {
-                    double z = camera.getTranslateZ();
-                    double newZ = z + mouseDeltaX * MOUSE_SPEED * modifier;
-                    camera.setTranslateZ(newZ);
-                }
-            }
-        });
     }
 }
